backend/requirements.txt
fastapi
uvicorn
pandas
google-api-python-client
youtube-transcript-api
requests
python-dotenv 